% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Simulated Example Data}
\format{
A data frame with 1500 rows and 11 variables where V1-V10 are the
predictors and V11 is the dependent variable.
}
\usage{
exampleData
}
\description{
Simulated data set with 1500 observations for illustrational
purposes.
}
\keyword{datasets}
